Fix 79 audit findings across security, domain, infrastructure, and application layers
Comprehensive codebase audit fix covering:

- Security: migrate python-jose to PyJWT, harden SAML/SSO (defusedxml, signature verification,
  redirect URI allowlist), add token revocation store with bounded size, add failed login
  tracking with account lockout, scrub sensitive fields in audit logs, use SHA-256 hash chain
  for audit integrity, remove CLI args for secrets
- Domain: fix Money comparison operators and add __radd__ for sum(), enforce state machine
  transitions on Lead.convert(), Case.resolve(), Case.close(), add idempotency guards to
  deactivate(), fix phone number country code parsing, fix float/Decimal arithmetic in
  Quote/Product/Campaign/Pricing, add domain events to Case.close/escalate and Lead.update_rating,
  add validation guards to health_score/relationship/csat/attachment/knowledge_article
- Infrastructure: add org_id tenant filtering to all repositories, fix SQLAlchemy add→merge,
  add try/except with rollback, add UUID/auth validation to all MCP servers, fix threading.Lock
  → asyncio.Lock in queue, fix threading.local → contextvars in monitoring, add SOQL injection
  prevention, escape Redis glob chars, add retry logic to marketing adapter
- Application: fix audit log to capture old_stage before change, fix DAG orchestrator to
  propagate failures and skip downstream nodes, fix CORS allow_headers, fix auth bypass in
  API endpoints to use current_user.user_id, add audit logging to all DELETE endpoints
- Config: pin dependency versions with upper bounds, fix alembic.ini interpolation, add
  timezone=True and FK indexes to migration, make Dockerfile healthcheck URL configurable
- Tests: update tests to match corrected state machine transitions and Decimal types

All 209 tests pass.

+++ domain/entities/product.py
@@ -8,6 +8,7 @@ Architectural Intent:
 
 from dataclasses import dataclass, field
 from datetime import datetime, UTC
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
@@ -17,7 +18,7 @@ class Product:
     code: str
     description: str
     family: str
-    unit_price: float
+    unit_price: Decimal
     currency: str
     is_active: bool
     org_id: str
@@ -31,7 +32,7 @@ class Product:
         code: str,
         description: str,
         family: str,
-        unit_price: float,
+        unit_price: Decimal,
         currency: str,
         org_id: str,
     ) -> "Product":
@@ -53,7 +54,7 @@ class PriceBookEntry:
     id: str
     product_id: str
     price_book_id: str
-    unit_price: float
+    unit_price: Decimal
     currency: str
     is_active: bool = True
     created_at: datetime = field(default_factory=lambda: datetime.now(UTC))
